fix: candidates were always undefined / type error on get
Fixes two issues:

1) tools.getCandidates.get always returned undefined[][] because of the
   wrong array iteration keyword

2) the return type on tools.getCandidates.get was wrong, saying it
   returned string[], but it actually returned string[][]

Also improves tools.getCandidates.get by changing the return type to use
true instead of boolean. This helps because you can now just say
`if (candidates)` instead of needing a type check. This is demonstrated
in the demo/test.ts file, which now calls getCandidates so that it's
easy to see if the fixes ever regress.